requirements.txt
# Production dependencies
psycopg2-binary==2.9.3
python-dotenv==0.20.0
structlog==22.1.0
uuid==1.30

# Development dependencies
pytest==7.1.2
pytest-cov==4.0.0
flake8==6.0.0
black==23.3.0
mypy==1.3.0
bandit==1.7.5
safety==2.3.4